APIs/atlassian.com/jira/1001.0.0-SNAPSHOT/python-aiohttp-client/tests/test_screen_tabs_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.screenable_tab import ScreenableTab


pytestmark = pytest.mark.asyncio

async def test_add_screen_tab(client):
    """Test case for add_screen_tab

    Create screen tab
    """
    body = {"name":"name","id":0}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='POST',
        path='/rest/api/3/screens/{screen_id}/tabs'.format(screen_id=56),
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_screen_tab(client):
    """Test case for delete_screen_tab

    Delete screen tab
    """
    headers = { 
        'Authorization': 'Bearer special-key',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='DELETE',
        path='/rest/api/3/screens/{screen_id}/tabs/{tab_id}'.format(screen_id=56, tab_id=56),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_all_screen_tabs(client):
    """Test case for get_all_screen_tabs

    Get all screen tabs
    """
    params = [('projectKey', 'project_key_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/rest/api/3/screens/{screen_id}/tabs'.format(screen_id=56),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_move_screen_tab(client):
    """Test case for move_screen_tab

    Move screen tab
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='POST',
        path='/rest/api/3/screens/{screen_id}/tabs/{tab_id}/move/{pos}'.format(screen_id=56, tab_id=56, pos=56),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_rename_screen_tab(client):
    """Test case for rename_screen_tab

    Update screen tab
    """
    body = {"name":"name","id":0}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='PUT',
        path='/rest/api/3/screens/{screen_id}/tabs/{tab_id}'.format(screen_id=56, tab_id=56),
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

